+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- A `Running` property has been added to Rumor
+
+### Changed
+- `Rumor.Run()` will restart the script instead of throwing an exception if it
+  is already running
+
 ## [0.1.0] - 2016-10-23
 Initial release.